+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+- Add debug options for sfdx-hardis commands
+- Add new sandbox refresh commands
+- New UI to update VsCode SFDX-Hardis configuration
+
 ## [6.0.9 (pre-release)] 2024-08-14
 
 - Add a new mode (by default) to execute sfdx-hardis commands in background, not in terminal